Project_B. New schedule system.

+++ Code/Project_B/CodeServerSide/DataProvider/DataHelper/ProcessData/ProcessStat.cs
@@ -1,12 +1,12 @@
 ﻿
 namespace Project_B.CodeServerSide.DataProvider.DataHelper.ProcessData {
     public class ProcessStat {
-        public int TotalCount;
-        public int MatchedCount;
-        public int CreateRawCount;
-        public int CreateOriginalCount;
-        public int FinallySuccessCount;
-        public int TryMatchedCount;
-        public int TryCreateOriginalCount;
+        public short TotalCount;
+        public short MatchedCount;
+        public short CreateRawCount;
+        public short CreateOriginalCount;
+        public short FinallySuccessCount;
+        public short TryMatchedCount;
+        public short TryCreateOriginalCount;
     }
 }
